fix: Prevent incorrect stopping condition for control loop
Defining the 'reached' variables as 'False' outside the while loop
of the control function caused the control loop to stop when the
desired states were achieved independently instead of simultaneously.
This resulted in incorrect control of the manipulator.

This was fixed by relocating the variable definition within the while
loop which ensured that the control loop would stop only if all the
three joint states for a given pose were achieved all at once.