fix: Prevent error propagation between poses
For the integral term in the control input, the error accumulates
over time for a given pose. When going from one pose to the next,
this error term should be made zero to start fresh for the new pose.
This update was missed which caused the accumulation to result in
incorrect control.

This was noticed because Ki value for 1st pose resulted in critical
control but as the manipulator moved towards the latter poses, the
integral term's effect far outweighed that of the proportional or
derivative terms.

This was corrected by updating the 'update_set_point' function to make
the integral term zero.