fix: Correct stopping condition on reaching individual poses
The requirement of having the manipulator stop for 1s after reaching
each pose was executed incorrectly where each joint was being stopped
for 1s as and when joints reached the desired position in the pose.

This was corrected by having the joints stop simultaneously after a
particular pose was reached. Also, instead of publishing zero control
input to freeze the joints, a negative control input equal in magnitude
to the last control input before reaching the pose was published to
oppose the motion due to inertia.